MVP Stats Scraper
I was unable to get this to work with the regular scraper and ended up using an NBA API endpoint to pull the mvp stats for each season. Basketballreference has some sort of automated request blocking on the individual player pages that was not an  issue while scraping just mvp votes. The new scraper reads our existing mvp voting data, and finds that players's stats from the season(s) that they recieved 100+ MVP voting points.